requirements.txt
Flask
gspread
oauth2client
Pillow
openai
